+++ src/nco/nco_rth_flt.c
@@ -1,4 +1,4 @@
-/* $Header: /data/zender/nco_20150216/nco/src/nco/nco_rth_flt.c,v 1.41 2015-01-20 18:20:15 zender Exp $ */
+/* $Header: /data/zender/nco_20150216/nco/src/nco/nco_rth_flt.c,v 1.42 2015-02-04 04:16:24 zender Exp $ */
 
 /* Purpose: Float-precision arithmetic, MSVC macros */
 
@@ -13,14 +13,14 @@
    MSVC does not define lrint(), lrintf(), lrintl(), llrint(), llrintf(), llrintl(): Round to nearest even integer, raise exceptions
    Summary of POSIX, ISO, and MSVC math intrinsics at http://www.johndcook.com/math_h.html */
 #ifdef _MSC_VER
-long long int llrint(double x){return (x >= 0.0) ? floor(x+0.5) : ceil(x-0.5);};
-long long int llrintf(float x){return (x >= 0.0f) ? floorf(x+0.5f) : ceilf(x-0.5f);};
-long int lrint(double x){return (x >= 0.0) ? floor(x+0.5) : ceil(x-0.5);};
-long int lrintf(float x){return (x >= 0.0f) ? floorf(x+0.5f) : ceilf(x-0.5f);};
-long long int llround(double x){return floor(x+0.5);}
-long long int llroundf(float x){return floorf(x+0.5f);}
-long int lround(double x){return floor(x+0.5);}
-long int lroundf(float x){return floorf(x+0.5f);}
+long long int llrint(double x){return (x >= 0.0) ? (long long int)floor(x+0.5) : (long long int)ceil(x-0.5);}; /* casts are CEWI for MSVC */
+long long int llrintf(float x){return (x >= 0.0f) ? (long long int)floorf(x+0.5f) : (long long int)ceilf(x-0.5f);};
+long int lrint(double x){return (x >= 0.0) ? (long int)floor(x+0.5) : (long int)ceil(x-0.5);};
+long int lrintf(float x){return (x >= 0.0f) ? (long int)floorf(x+0.5f) : (long int)ceilf(x-0.5f);};
+long long int llround(double x){return (long long int)floor(x+0.5);}
+long long int llroundf(float x){return (long long int)floorf(x+0.5f);}
+long int lround(double x){return (long int)floor(x+0.5);}
+long int lroundf(float x){return (long int)floorf(x+0.5f);}
 #endif /* !_MSC_VER */ 
 
 /* In ANSI C, <math.h> provides standard math intrinsics in double precision
